+++ Kernel/Output/HTML/Preferences/CustomService.pm
@@ -17,6 +17,7 @@ our @ObjectDependencies = (
     'Kernel::System::DB',
     'Kernel::System::Service',
     'Kernel::System::Web::Request',
+    'Kernel::Config',
 );
 
 sub new {
@@ -39,8 +40,14 @@ sub Param {
     my @Params;
     my @CustomServiceIDs;
 
-    # check needed param, if no user id is given, do not show this box
-    if ( !$Param{UserData}->{UserID} ) {
+    # check needed param,
+    # if no user id is given or Ticket::Service is disabled
+    # do not show this box
+    if (
+        !$Param{UserData}->{UserID}
+        || !$Kernel::OM->Get('Kernel::Config')->Get('Ticket::Service')
+        )
+    {
         return ();
     }